Remove default rest easy http headers and allow custom http headers
